+++ Sources/Sentry/SentrySwizzle.m
@@ -21,9 +21,7 @@ SentrySwizzleInfo ()
 {
     NSAssert(_impProviderBlock, nil);
 
-#if TEST
     self.originalCalled = true;
-#endif
 
     // Casting IMP to SentrySwizzleOriginalIMP to force user casting.
     return (SentrySwizzleOriginalIMP)_impProviderBlock();
